Fix consistency checks

+++ chocolate-doom/src/g_game.c
@@ -253,7 +253,7 @@ boolean         precache = true;        // if true, load all graphics at start
  
 wbstartstruct_t wminfo;               	// parms for world map / intermission 
  
-short		consistancy[MAXPLAYERS][BACKUPTICS]; 
+byte		consistancy[MAXPLAYERS][BACKUPTICS]; 
  
  
 // 
@@ -850,14 +850,12 @@ void G_Ticker (void)
 			
 	    if (netgame && !netdemo && !(gametic%ticdup) ) 
 	    { 
-                /*
 		if (gametic > BACKUPTICS 
 		    && consistancy[i][buf] != cmd->consistancy) 
 		{ 
 		    I_Error ("consistency failure (%i should be %i)",
 			     cmd->consistancy, consistancy[i][buf]); 
 		} 
-                */
 		if (players[i].mo) 
 		    consistancy[i][buf] = players[i].mo->x; 
 		else
